pyproject.toml
[project]
name = "image-consolidation"
version = "0.1.0"
description = "Clean up and organize image/video libraries at scale (~1TB)"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.4",
    "exifread>=3.0",
    "imagehash>=4.3",
    "faiss-cpu>=1.8",
    "typer>=0.12",
    "tqdm>=4.66",
    "pydantic>=2.7",
    "rich>=13.7",
]

[project.scripts]
imgc = "image_consolidation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_consolidation"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
